requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
supabase>=1.0.0
python-dotenv>=0.21.0
pandas
scikit-learn